requirements.txt
numpy==1.21.5
SimpleITK==2.1.1.2
tqdm==4.64.0
elasticdeform==0.4.9
matplotlib==3.5.2
scipy==1.8.0
scikit-image==0.19.2
wandb==0.12.16
albumentations==1.1.0
pandas==1.4.2
patool==1.12
pydicom==2.3.0